Cleaned up logDetect, testing on nano
