--- Computer-Vision/eventLogs.csv
@@ -1,79 +1,0 @@
-ID,Timestamp,Process ID,Tag,Module,Level,Message
-1,2024-03-08 12:18:13.287,,get transceiver,detector,1,A test to see if get transceiver reports
-2,2024-03-08 12:18:49.720,,get transceiver,detector,1,A test to see if get transceiver reports
-3,2024-03-08 12:20:07.335,,get transceiver,detector,1,A test to see if get transceiver reports
-4,2024-03-08 12:26:59.293,,get transceiver,detector,1,A test to see if get transceiver reports
-5,2024-03-12 16:27:26.122,,get transceiver,detector,1,A test to see if get transceiver reports
-6,2024-03-12 20:28:36,,get transceiver,detector,1,A test to see if get transceiver reports
-7,2024-03-12 16:49:08.252,,get transceiver,detector,1,A test to see if get transceiver reports
-8,2024-03-12 16:51:27.676,,get transceiver,detector,1,A test to see if get transceiver reports
-9,2024-03-12 17:05:26.924,,get transceiver,detector,1,A test to see if get transceiver reports
-10,2024-03-12 17:16:27.873,,get transceiver,detector,1,A test to see if get transceiver reports
-11,2024-03-12 17:19:21.961,,get transceiver,detector,1,A test to see if get transceiver reports
-12,2024-03-12 17:20:08.528,,get transceiver,detector,1,A test to see if get transceiver reports
-13,2024-03-12 17:26:43.760,,get transceiver,detector,1,A test to see if get transceiver reports
-14,2024-03-12 17:28:12.018,,get transceiver,detector,1,A test to see if get transceiver reports
-15,2024-03-12 17:32:04.470,,get transceiver,detector,1,A test to see if get transceiver reports
-16,2024-03-12 17:33:26.777,,get transceiver,detector,1,A test to see if get transceiver reports
-17,2024-03-12 17:36:16.707,,get transceiver,detector,1,A test to see if get transceiver reports
-18,2024-03-12 17:38:12.364,,get transceiver,detector,1,A test to see if get transceiver reports
-19,2024-03-12 17:44:43.922,,get transceiver,detector,1,A test to see if get transceiver reports
-20,2024-03-12 17:53:39.587,,get transceiver,detector,1,A test to see if get transceiver reports
-21,2024-03-12 17:53:51.967,,get transceiver,detector,1,A test to see if get transceiver reports
-22,2024-03-13 02:59:14.466,,get transceiver,detector,1,A test to see if get transceiver reports
-23,2024-03-13 03:02:41.521,,get transceiver,detector,1,A test to see if get transceiver reports
-24,2024-03-13 03:05:20.451,,get transceiver,detector,1,A test to see if get transceiver reports
-25,2024-03-13 03:08:37.896,,get transceiver,detector,1,A test to see if get transceiver reports
-26,2024-03-13 03:12:56.073,,get transceiver,detector,1,A test to see if get transceiver reports
-27,2024-03-13 03:14:03.831,,get transceiver,detector,1,A test to see if get transceiver reports
-28,2024-03-13 03:16:06.327,,get transceiver,detector,1,A test to see if get transceiver reports
-29,2024-03-13 03:16:33.652,,get transceiver,detector,1,A test to see if get transceiver reports
-30,2024-03-13 03:16:55.299,,get transceiver,detector,1,A test to see if get transceiver reports
-31,2024-03-13 03:17:46.463,,get transceiver,detector,1,A test to see if get transceiver reports
-32,2024-03-13 03:19:22.624,,get transceiver,detector,1,A test to see if get transceiver reports
-33,2024-03-13 03:20:48.612,,get transceiver,detector,1,A test to see if get transceiver reports
-34,2024-03-13 03:24:49.853,,get transceiver,detector,1,A test to see if get transceiver reports
-35,2024-03-13 03:26:44.635,,get transceiver,detector,1,A test to see if get transceiver reports
-36,2024-03-13 03:28:51.086,,get transceiver,detector,1,A test to see if get transceiver reports
-37,2024-03-13 03:34:29.710,,get transceiver,detector,1,A test to see if get transceiver reports
-38,2024-03-13 03:36:36.582,,get transceiver,detector,1,A test to see if get transceiver reports
-39,2024-03-13 03:43:56.171,,get transceiver,detector,1,A test to see if get transceiver reports
-40,2024-03-13 03:45:31.926,,get transceiver,detector,1,A test to see if get transceiver reports
-41,2024-03-13 03:53:19.684,,get transceiver,detector,1,A test to see if get transceiver reports
-42,2024-03-13 03:55:13.401,,get transceiver,detector,1,A test to see if get transceiver reports
-43,2024-03-13 03:57:18.240,,get transceiver,detector,1,A test to see if get transceiver reports
-44,2024-03-13 04:01:42.866,,get transceiver,detector,1,A test to see if get transceiver reports
-45,2024-03-13 04:05:43.382,,get transceiver,detector,1,A test to see if get transceiver reports
-46,2024-03-13 04:06:02.680,,get transceiver,detector,1,A test to see if get transceiver reports
-47,2024-03-13 04:12:37.078,,get transceiver,detector,1,A test to see if get transceiver reports
-48,2024-03-13 04:16:30.907,,get transceiver,detector,1,A test to see if get transceiver reports
-49,2024-03-13 04:18:06.250,,get transceiver,detector,1,A test to see if get transceiver reports
-50,2024-03-13 04:19:00.772,,get transceiver,detector,1,A test to see if get transceiver reports
-51,2024-03-14 22:47:08.324,,get transceiver,detector,1,A test to see if get transceiver reports
-52,2024-03-14 22:48:16.045,,get transceiver,detector,1,A test to see if get transceiver reports
-53,2024-03-14 22:50:44.954,,get transceiver,detector,1,A test to see if get transceiver reports
-54,2024-03-14 22:52:09.952,,get transceiver,detector,1,A test to see if get transceiver reports
-55,2024-03-14 22:52:37.435,,get transceiver,detector,1,A test to see if get transceiver reports
-56,2024-03-14 22:58:58.831,,get transceiver,detector,1,A test to see if get transceiver reports
-57,2024-03-14 22:59:12.938,,get transceiver,detector,1,A test to see if get transceiver reports
-58,2024-03-14 23:02:17.477,,get transceiver,detector,1,A test to see if get transceiver reports
-59,2024-03-14 23:02:36.155,,get transceiver,detector,1,A test to see if get transceiver reports
-60,2024-03-14 23:04:43.819,,get transceiver,detector,1,A test to see if get transceiver reports
-61,2024-03-14 23:05:20.723,,get transceiver,detector,1,A test to see if get transceiver reports
-62,2024-03-14 23:05:40.687,,get transceiver,detector,1,A test to see if get transceiver reports
-63,2024-03-14 23:06:35.527,,get transceiver,detector,1,A test to see if get transceiver reports
-64,2024-03-14 23:07:24.734,,get transceiver,detector,1,A test to see if get transceiver reports
-65,2024-03-15 11:17:15.225,,get transceiver,detector,1,A test to see if get transceiver reports
-66,2024-03-25 11:09:29.931,,get transceiver,detector,1,A test to see if get transceiver reports
-67,2024-03-25 11:16:07.547,,get transceiver,detector,1,A test to see if get transceiver reports
-68,2024-03-25 11:18:00.535,,get transceiver,detector,1,A test to see if get transceiver reports
-69,2024-03-25 11:18:07.538,,get transceiver,detector,1,A test to see if get transceiver reports
-70,2024-03-25 11:35:13.824,,get transceiver,detector,1,A test to see if get transceiver reports
-71,2024-03-25 12:08:08.328,,get transceiver,detector,1,A test to see if get transceiver reports
-72,2024-03-25 12:12:49.691,,get transceiver,detector,1,A test to see if get transceiver reports
-73,2024-03-25 12:16:51.297,,get transceiver,detector,1,A test to see if get transceiver reports
-74,2024-03-25 12:47:28.932,,get transceiver,detector,1,A test to see if get transceiver reports
-75,2024-03-25 13:11:59.630,,get transceiver,detector,1,A test to see if get transceiver reports
-76,2024-03-25 14:38:10.258,,get transceiver,detector,1,A test to see if get transceiver reports
-77,2024-03-25 14:51:20.688,,get transceiver,detector,1,A test to see if get transceiver reports
-78,2024-03-25 15:08:43.229,,get transceiver,detector,1,A test to see if get transceiver reports
