--- Computer-Vision/testLogger.py
@@ -1,7 +1,0 @@
-from Logger import Logger
-
-logger = Logger()
-
-# Make a test log
-logger.addEvent('get transceiver', 'detector', '1', 'A test to see if get transceiver reports')
-logger.addData ()
